+++ inst/GerminaQuant/ui.R
@@ -34,7 +34,7 @@ shinyUI(dashboardPage(skin = "green",
         menuItem("Exploratory", tabName = "outlier", icon = icon("search")),
         menuItem("Statistics", tabName = "stat", icon = icon("pie-chart")),
         menuItem("Graphics", tabName = "graph", icon = icon("tint")),
-        menuItem("Intime", tabName = "germint", icon = icon("hourglass")),
+        menuItem("InTime", tabName = "germint", icon = icon("hourglass")),
         menuItem("Tools", tabName = "tools", icon = icon("wrench"))
       )
 
@@ -168,7 +168,7 @@ shinyUI(dashboardPage(skin = "green",
                   
                       <div style="display:inline-block; width:32%">
                       <p style="text-align:center">
-                      <a target="_blank" href="https://flavjack.shinyapps.io/germinaquant/"><img src="https://flavjack.github.io/GerminaR/reference/figures/germinaquant.png" style="height:70px" title="GerminaQuant" alt="GerminaQuant for R"></a>
+                      <a target="_blank" href="https://flavjack.shinyapps.io/germinaquant/"><img src="https://flavjack.github.io/GerminaR/img/germinaquant.png" style="height:80px" title="GerminaQuant" alt="GerminaQuant for R"></a>
                       <span style="display:block;"><small>GerminaQuant</small></span>
                       </p></div>
                       
@@ -262,16 +262,8 @@ shinyUI(dashboardPage(skin = "green",
 
           textInput("SeedN", label = strong("Seeds (column name)"), value = "seeds"),
           
-          textInput("evalName", label = strong("Evaluations (prefix)"), value = "D"),
+          textInput("evalName", label = strong("Evaluations (prefix)"), value = "D")
           
-          uiOutput("filter_01"),
-          
-          uiOutput("filter_fact01"),
-          
-          uiOutput("filter_02"),
-          
-          uiOutput("filter_fact02")
-
           )
           
         )
@@ -286,7 +278,14 @@ shinyUI(dashboardPage(skin = "green",
           
           box(width = 12,
               
-              DT::dataTableOutput("summary")
+              DT::dataTableOutput("summary"),
+              
+              br(),
+              
+              actionButton(inputId= 'indicesLink', label = "Indices info"
+                           , icon = icon("seedling")
+                           , onclick ="window.open('https://germinar.inkaverse.com/articles/introduction.html', '_blank')"
+                           )
               
           )